SOC-2637 | reported filter to the left side
