+++ front/main/app/components/discussion-entities-list-layout.js
@@ -1,5 +1,3 @@
 import DiscussionModalDialogMixin from '../mixins/discussion-modal-dialog';
 
-export default Ember.Component.extend(DiscussionModalDialogMixin, {
-	discussionSort: Ember.inject.service(),
-});
+export default Ember.Component.extend(DiscussionModalDialogMixin, {});
