+++ front/main/app/templates/components/discussion-entities-list-layout.hbs
@@ -38,15 +38,14 @@
 					}}
 				{{/if}}
 			{{/if}}
+			{{#if model.canModerate}}
+				{{discussion-reported-filter}}
+			{{/if}}
 		</div>
 		<div class="small-12 large-6 columns">
 			{{yield}}
 		</div>
 		<div class="rail right large-3 columns">
-			{{discussion-filters
-				applyFilters=(action this.attrs.applyFilters)
-				canModerate=model.canModerate
-			}}
 			{{discussion-community-unit
 				displayGuidelinesLink=true
 				displayWikiaHomeLink=true
