+++ front/main/app/templates/components/discussion-filters.hbs
@@ -26,16 +26,18 @@
 				</ul>
 			</fieldset>
 		{{/if}}
-		{{#if canModerate}}
-			<fieldset class="discussion-fieldset moderation-fieldset">
-				<legend>{{i18n 'main.moderation-filters-header' ns='discussion'}}</legend>
-				{{wikia-ui-components/fancy-check-box
-					class='moderation-checkbox-label'
-					isChecked=onlyReported
-					label=(i18n 'main.reported-content-filter-text' ns='discussion')
-					click=(action 'toggleReported')
-				}}
-			</fieldset>
+		{{#if showReportedFilter}}
+			{{#if canModerate}}
+				<fieldset class="discussion-fieldset moderation-fieldset">
+					<legend>{{i18n 'main.moderation-filters-header' ns='discussion'}}</legend>
+					{{wikia-ui-components/fancy-check-box
+						class='moderation-checkbox-label'
+						isChecked=onlyReported
+						label=(i18n 'main.reported-content-filter-text' ns='discussion')
+						click=(action 'toggleReported')
+					}}
+				</fieldset>
+			{{/if}}
 		{{/if}}
 	</div>
 	{{#if showApplyButton}}
