+++ front/main/app/templates/components/discussion-header.hbs
@@ -27,6 +27,7 @@
 					categories=categories
 					showApplyButton=true
 					showCategories=showCategories
+					showReportedFilter=true
 					showSortSection=true
 				}}
 			{{/pop-over}}
