+++ front/main/app/templates/components/discussion-reported-filter.hbs
@@ -1,0 +1,7 @@
+<legend>{{i18n 'main.moderation-filters-header' ns='discussion'}}</legend>
+{{wikia-ui-components/fancy-check-box
+	class='moderation-checkbox-label'
+	isChecked=onlyReported
+	label=(i18n 'main.reported-content-filter-text' ns='discussion')
+	click=(action 'toggleReported')
+}}
